removed old test projects and updated sln file
