+++ PaginableCollections.AspNetCore.sln
@@ -1,4 +1,4 @@
-﻿
+
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26730.16
@@ -12,34 +12,74 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "build", "build", "{288485C6
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PaginableCollections.AspNetCore", "src\PaginableCollections.AspNetCore\PaginableCollections.AspNetCore.csproj", "{CFBF66A6-08AE-4DA1-92F4-18E24CC793B1}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{FE767F3D-FD81-459E-9BD0-1B9130CE14F8}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{64EBDAF1-8B29-404C-AD2C-C4D6E9FBCE4E}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{408B4CEC-0F1A-4DF5-8DFC-BE31149C94B5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaginableCollections.AspNetCore.IntegrationTests.Condensed", "tests\PaginableCollections.AspNetCore.IntegrationTests.Condensed\PaginableCollections.AspNetCore.IntegrationTests.Condensed.csproj", "{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PaginableCollections.AspNetCore.IntegrationTests", "tests\PaginableCollections.AspNetCore.IntegrationTests\PaginableCollections.AspNetCore.IntegrationTests.csproj", "{C44479EA-B953-49E9-A240-7477571B5FF9}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaginableCollections.AspNetCore.IntegrationTests.Expanded", "tests\PaginableCollections.AspNetCore.IntegrationTests.Expanded\PaginableCollections.AspNetCore.IntegrationTests.Expanded.csproj", "{2199596C-5BF8-446E-A6C0-B8F74326C68C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaginableCollections.AspNetCore.IntegrationTests.LinkBased", "tests\PaginableCollections.AspNetCore.IntegrationTests.LinkBased\PaginableCollections.AspNetCore.IntegrationTests.LinkBased.csproj", "{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{CFBF66A6-08AE-4DA1-92F4-18E24CC793B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CFBF66A6-08AE-4DA1-92F4-18E24CC793B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CFBF66A6-08AE-4DA1-92F4-18E24CC793B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CFBF66A6-08AE-4DA1-92F4-18E24CC793B1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C44479EA-B953-49E9-A240-7477571B5FF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C44479EA-B953-49E9-A240-7477571B5FF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C44479EA-B953-49E9-A240-7477571B5FF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C44479EA-B953-49E9-A240-7477571B5FF9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Debug|x64.ActiveCfg = Debug|x64
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Debug|x64.Build.0 = Debug|x64
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Debug|x86.ActiveCfg = Debug|x86
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Debug|x86.Build.0 = Debug|x86
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Release|x64.ActiveCfg = Release|x64
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Release|x64.Build.0 = Release|x64
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Release|x86.ActiveCfg = Release|x86
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA}.Release|x86.Build.0 = Release|x86
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Debug|x64.ActiveCfg = Debug|x64
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Debug|x64.Build.0 = Debug|x64
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Debug|x86.ActiveCfg = Debug|x86
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Debug|x86.Build.0 = Debug|x86
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Release|x64.ActiveCfg = Release|x64
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Release|x64.Build.0 = Release|x64
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Release|x86.ActiveCfg = Release|x86
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C}.Release|x86.Build.0 = Release|x86
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Debug|x64.ActiveCfg = Debug|x64
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Debug|x64.Build.0 = Debug|x64
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Debug|x86.ActiveCfg = Debug|x86
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Debug|x86.Build.0 = Debug|x86
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Release|x64.ActiveCfg = Release|x64
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Release|x64.Build.0 = Release|x64
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Release|x86.ActiveCfg = Release|x86
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{C44479EA-B953-49E9-A240-7477571B5FF9} = {FE767F3D-FD81-459E-9BD0-1B9130CE14F8}
-	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {ACF63DCD-C48C-46E0-BA66-DC714740C160}
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{4FCD9946-B5A6-4984-BB60-8A73E0C354AA} = {64EBDAF1-8B29-404C-AD2C-C4D6E9FBCE4E}
+		{2199596C-5BF8-446E-A6C0-B8F74326C68C} = {64EBDAF1-8B29-404C-AD2C-C4D6E9FBCE4E}
+		{04A0E98D-9F99-4D86-896D-D7D4E6391ADF} = {64EBDAF1-8B29-404C-AD2C-C4D6E9FBCE4E}
 	EndGlobalSection
 EndGlobal
