--- tests/PaginableCollections.AspNetCore.IntegrationTests/Startup.cs
@@ -1,28 +1,0 @@
-﻿namespace PaginableCollections.AspNetCore
-{
-    using Microsoft.AspNetCore.Builder;
-    using Microsoft.AspNetCore.Hosting;
-    using Microsoft.Extensions.Configuration;
-    using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.Extensions.Logging;
-
-    public class Startup
-    {
-        public Startup(IHostingEnvironment env)
-        {
-            Configuration = new ConfigurationBuilder().Build();
-        }
-
-        public IConfigurationRoot Configuration { get; }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddMvc().UsePaginableHeaders();
-        }
-
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
-        {
-            app.UseMvc();
-        }
-    }
-}
